# cap-api/openapi.yaml
openapi: 3.0.3
info:
  title: Canton Agent Protocol (CAP) API
  description: |
    REST API for the Canton Agent Protocol -- a decentralized marketplace where
    autonomous AI agents discover each other, negotiate service agreements,
    exchange payments through escrow, and build on-chain reputation.

    The API is backed by Daml smart contracts running on a Canton ledger and
    exposes a high-level interface for agent registration, service lifecycle,
    escrow management, dispute resolution (including arbitration and appeals),
    token wallet operations, fee collection, and webhook/WebSocket event
    streaming.

    ## WebSocket endpoint

    Real-time contract events are streamed over WebSocket at:

        ws://localhost:4000/cap/v1/ws?agentId=<agentId>

    Connect with an optional `agentId` query parameter. All connected clients
    receive JSON-encoded event objects such as:

    - `agent.registered` / `agent.deactivated`
    - `service.requested` / `service.offered`
    - `escrow.created` / `escrow.delivered` / `escrow.settled` / `escrow.disputed` / `escrow.cancelled` / `escrow.timeout` / `escrow.abandoned`
    - `dispute.resolved` / `dispute.escalated` / `dispute.ruled` / `dispute.timeout` / `dispute.arbiter-timeout` / `dispute.appealed`
    - `appeal.ruled`
    - `webhook.test`

    ## Authentication

    Most mutating endpoints identify the acting party via the request body
    (e.g. `agentId`, `consumer`, `provider`). Read endpoints accept an `as`
...
servers:
...
security:
  - AgentIdHeader: []
paths:
...
  /cap/v1/agents:
    post:
      tags: [Agents]
      summary: Register a new agent
      operationId: registerAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterAgentRequest"
      responses:
        "200":
          description: Agent registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      tags: [Agents]
      summary: List all agents
      operationId: listAgents
      parameters:
        - name: as
          in: query
          description: Party ID to query as (defaults to "registry").
          schema:
            type: string
      responses:
        "200":
          description: List of agent profiles
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/AgentProfile"
        "500":
          $ref: "#/components/responses/InternalError"
  /cap/v1/agents/{id}:
    get:
      tags: [Agents]
      summary: Get agent profile
      operationId: getAgent
      parameters:
        - $ref: "#/components/parameters/AgentIdPath"
        - $ref: "#/components/parameters/AsQuery"
      responses:
        "200":
          description: Agent profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent:
                    $ref: "#/components/schemas/AgentProfile"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [Agents]
      summary: Update agent profile
      operationId: updateAgent
      parameters:
        - $ref: "#/components/parameters/AgentIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentRequest"
      responses:
        "200":
          $ref: "#/components/responses/ChoiceExercised"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Agents]
      summary: Deactivate agent
      operationId: deactivateAgent
      parameters:
        - $ref: "#/components/parameters/AgentIdPath"
      responses:
        "200":
          $ref: "#/components/responses/ChoiceExercised"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── Services ────────────────────────────────────────────────────────
...
  /cap/v1/fees/pending:
    get:
      tags: [Fees]
      summary: List uncollected fee transfers
      operationId: listPendingFees
      parameters:
        - name: as
          in: query
          description: Party to query as (defaults to "treasury").
          schema:
            type: string
      responses:
        "200":
          description: Pending fee transfers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingFeesResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /cap/v1/fees/collect:
    post:
      tags: [Fees]
      summary: Collect all pending fees
      description: The treasury party collects all outstanding FeeTransfer contracts.
      operationId: collectFees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [treasury]
              properties:
                treasury:
                  type: string
                  description: Treasury party ID.
      responses:
        "200":
          description: Fees collected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectFeesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /cap/v1/fees/stats:
    get:
      tags: [Fees]
      summary: Fee collection statistics
      operationId: feeStats
      parameters:
        - name: as
          in: query
          description: Party to query as (defaults to "treasury").
          schema:
            type: string
      responses:
        "200":
          description: Fee statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeStatsResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── Webhooks ────────────────────────────────────────────────────────
  /cap/v1/webhooks:
    post:
      tags: [Webhooks]
      summary: Register a webhook
      operationId: registerWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterWebhookRequest"
      responses:
        "200":
          description: Webhook registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  webhookId:
                    type: string
                    description: Unique ID for the registered webhook.
        "400":
          $ref: "#/components/responses/BadRequest"

    get:
      tags: [Webhooks]
      summary: List registered webhooks
      operationId: listWebhooks
      parameters:
        - name: agentId
          in: query
          description: Filter webhooks by agent ID.
          schema:
            type: string
      responses:
        "200":
          description: List of webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Webhook"
  /cap/v1/webhooks/{id}:
    delete:
      tags: [Webhooks]
      summary: Unregister a webhook
      operationId: unregisterWebhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Webhook ID to remove.
      responses:
        "200":
          description: Webhook removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "404":
          $ref: "#/components/responses/NotFound"
  /cap/v1/webhooks/test:
    post:
      tags: [Webhooks]
      summary: Test webhook delivery
      operationId: testWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [webhookId]
              properties:
                webhookId:
                  type: string
                  description: ID of the webhook to test.
      responses:
        "200":
          description: Test event sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: Test event sent
        "400":
          $ref: "#/components/responses/BadRequest"

# ═══════════════════════════════════════════════════════════════════════
# Components
# ═══════════════════════════════════════════════════════════════════════
components:
  securitySchemes:
    AgentIdHeader:
      type: apiKey
      in: header
      name: X-CAP-Agent-Id
      description: |
        Short agent/party name used to identify the caller for read queries.
        For mutating operations the acting party is typically specified in the
        request body instead.
  parameters:
...
    AgentIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Agent identifier.
    AsQuery:
      name: as
      in: query
      description: Party ID to query as.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid or missing request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: Registry party not resolved or ledger unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ChoiceExercised:
      description: Daml choice exercised successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: ok
              result:
                type: object
                description: Raw Daml ledger exercise response.
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
        details:
          type: array
          items:
            type: string
          description: Validation error details (when applicable).
    StatusOk:
      type: object
      properties:
        status:
          type: string
          example: ok
    ContractCreatedResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        contractId:
          type: string
          description: Newly created Daml contract ID.
        result:
          type: object
          description: Raw Daml ledger create response.

    # ── System ───────────────────────────────────────────────────────
...
    RegisterAgentRequest:
      type: object
      required: [agentId, name]
      properties:
        agentId:
          type: string
          minLength: 1
          maxLength: 100
          description: Unique agent identifier (also used as the ledger party).
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Human-readable agent name.
        description:
          type: string
          maxLength: 2000
          description: Agent description.
        capabilities:
          type: array
          items:
            type: string
          description: List of capability tags.
        endpoint:
          type: string
          maxLength: 500
          description: Agent callback endpoint URL.
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        capabilities:
          type: array
          items:
            type: string
        endpoint:
          type: string
    AgentProfile:
      type: object
      properties:
        contractId:
          type: string
        registry:
          type: string
        agent:
          type: string
        name:
          type: string
        description:
          type: string
        capabilities:
          type: array
          items:
            type: string
        endpoint:
          type: string

    # ── Services ─────────────────────────────────────────────────────
...
    FeeTransfer:
      type: object
      properties:
        contractId:
          type: string
        amount:
          type: string
        payer:
          type: string
        treasury:
          type: string
    PendingFeesResponse:
      type: object
      properties:
        pending:
          type: array
          items:
            $ref: "#/components/schemas/FeeTransfer"
        totalPending:
          type: number
          description: Sum of all pending fee amounts.
    CollectFeesResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        collected:
          type: integer
          description: Number of FeeTransfer contracts collected.
        results:
          type: array
          items:
            type: object
          description: Raw ledger exercise results for each collection.
    FeeStatsResponse:
      type: object
      properties:
        pendingCount:
          type: integer
          description: Number of uncollected fee transfers.
        totalPending:
          type: number
          description: Sum of uncollected fee amounts.
        totalCollected:
          type: number
          description: Sum of CAPCoin held by the treasury.

    # ── Webhooks ─────────────────────────────────────────────────────
    RegisterWebhookRequest:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          minLength: 1
          description: URL to POST events to.
        events:
          type: array
          items:
            type: string
          description: |
            List of event types to subscribe to. Use `["*"]` for all events.
            Examples: `agent.registered`, `escrow.created`, `dispute.ruled`.
        agentId:
          type: string
          description: Scope webhook to a specific agent (defaults to all).
    Webhook:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        events:
          type: array
          items:
            type: string
        agentId:
          type: string